pyproject.toml
[tool.poetry]
name = "info-about-book-api"
version = "0.1.0"
description = "API для управления информацией о книгах, авторах и тегах."
authors = ["Dmitriy Kultyshev <@kuldm>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.112.3"
sqlalchemy = "^2.0.34"
pydantic = "^2.8.2"
alembic = "^1.13.2"
uvicorn = "^0.30.6"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.4.0"
asyncpg = "^0.29.0"
passlib = "^1.7.4"
python-jose = "^3.3.0"
bcrypt = "^4.2.0"
loguru = "^0.7.2"
pytest = "^8.3.2"
pytest-asyncio = "^0.24.0"
httpx = "^0.27.2"
gunicorn = "^23.0.0"
elasticsearch = "7.10.1"
elasticsearch-dsl = "7.3.0"


[tool.poetry.group.dev.dependencies]
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
